backend/requirements.txt
pyspark==3.5.4
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
kafka-python-ng==2.2.3
pydantic-settings==2.6.1
